% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berkowitz.R
\docType{data}
\name{Berkowitz}
\alias{Berkowitz}
\title{Berkowitz Sight-Singing Database}
\format{
A function containing different versions of the Berkowitz data frame, returned with "main", "ngram", "phrases" and "files" keys
\describe{
  \item{melody}{price, in US dollars}
  \item{durations}{weight of the diamond, in carats}
  ...
}
}
\source{
\url{DOI:10.2307/3389405}
}
\usage{
Berkowitz
}
\description{
A dataset containing just over x melodies designed to improve sight-singing
}
\keyword{datasets}
