% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berkowitz.R
\docType{data}
\name{Berkowitz_subset}
\alias{Berkowitz_subset}
\title{A subset of Berkowitz optimised to help audio transcription}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}, \code{item_bank}) with 81254 rows and 44 columns.
}
\usage{
Berkowitz_subset
}
\description{
A subset of Berkowitz optimised to help audio transcription
}
\keyword{datasets}
