% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berkowitz.R
\docType{data}
\name{combined_item_bank}
\alias{combined_item_bank}
\title{Berkowitz Sight-Singing Item Bank (combined)}
\format{
Dataframe.
}
\source{
\url{DOI:10.2307/3389405}
}
\usage{
combined_item_bank
}
\description{
A dataset containing ~630 melodies designed to improve sight-singing (combined N-grams and phrases)
}
\keyword{datasets}
