% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berkowitz.R
\docType{data}
\name{item_item_bank}
\alias{item_item_bank}
\title{Berkowitz Sight-Singing Database (item version only)}
\format{
Dataframe.
}
\source{
\url{DOI:10.2307/3389405}
}
\usage{
item_item_bank
}
\description{
A dataset containing ~630 melodies designed to improve sight-singing
}
\keyword{datasets}
