% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Berkowitz.R
\docType{data}
\name{ngram_item_bank}
\alias{ngram_item_bank}
\title{Berkowitz Sight-Singing Item Bank (N-gram only)}
\format{
Dataframe.
}
\source{
\url{DOI:10.2307/3389405}
}
\usage{
ngram_item_bank
}
\description{
A dataset containing ~630 melodies designed to improve sight-singing (N-grams only)
}
\keyword{datasets}
