% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAA_models.R
\name{predict_arrhythmic_difficulty}
\alias{predict_arrhythmic_difficulty}
\title{Predict arrhythmic difficulty scores}
\usage{
predict_arrhythmic_difficulty(data)
}
\arguments{
\item{data}{}
}
\description{
Predict arrhythmic difficulty scores
}
