% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAA_models.R
\name{predict_rhythmic_difficulty}
\alias{predict_rhythmic_difficulty}
\title{Predict rhythmic difficulty scores}
\usage{
predict_rhythmic_difficulty(data)
}
\arguments{
\item{data}{}
}
\description{
Predict rhythmic difficulty scores
}
